+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,8 +1,16 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
+import kitchenpos.application.FakeKitchenridersClient;
+import kitchenpos.order.domain.InMemoryOrderRepository;
+import kitchenpos.application.InMemoryOrderTableRepository;
 import kitchenpos.domain.*;
 import kitchenpos.menu.domain.InMemoryMenuRepository;
 import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
